+++ Assets/Anim/AnimatorControllers/Player_Viewmodel.controller
@@ -3124,7 +3124,7 @@ AnimatorState:
   m_MirrorParameterActive: 0
   m_CycleOffsetParameterActive: 0
   m_Motion: {fileID: 7400046, guid: a89d22e0a3f90c549a3dabf775e7debf, type: 3}
-  m_Tag: 
+  m_Tag: magicAttack
   m_SpeedParameter: 
   m_MirrorParameter: 
   m_CycleOffsetParameter:
