+++ Assets/C#/WeaponScripts/BeamMagic.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: cf3b2078326bde6449a556f801cdf3a2
-timeCreated: 1502677016
+guid: 3d188e4bf4add466c89c390f8137cf38
+timeCreated: 1488067227
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
