+++ Assets/C#/WeaponScripts/BurstMagic.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: cf3b2078326bde6449a556f801cdf3a2
-timeCreated: 1502677016
+guid: 61b6c000cebb7f340aa76999ed0e2c97
+timeCreated: 1502677499
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
