+++ Assets/C#/WeaponScripts/SpawnerMagic.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: cf3b2078326bde6449a556f801cdf3a2
-timeCreated: 1502677016
+guid: 62655cb090cf0134da8155eb68f721dd
+timeCreated: 1502678191
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
