+++ Assets/Prefabs/Effects.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 180665b674d295148876e6731f68acc0
+folderAsset: yes
+timeCreated: 1502682962
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
