+++ Assets/Prefabs/Effects/SpawnParticles.prefab.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 3cd2cd8ec1d80c1498b580259f391b00
+timeCreated: 1502682968
+licenseType: Free
+NativeFormatImporter:
+  mainObjectFileID: 100100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
