+++ Assets/Prefabs/Pickups/Weapons/BoxSpawnerMagic.prefab.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 22c8d22aa2ae8cf4a8944f908caf40ab
+timeCreated: 1502678671
+licenseType: Free
+NativeFormatImporter:
+  mainObjectFileID: 100100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
